+++ easymode/admin/models/__init__.py
@@ -1,0 +1,3 @@
+"""
+General admin functionality related to models.
+"""